refactor: Adjust the role of index in collection
Now, the concept IndexProfile refers to the pure index profiles that
can be shared among collections, while the concept
CollectionsOnIndexProfiles refer to the instances of the pure index
profiles on each collection, respectively. Concretely, each collection
should have at most one instance for each pure index profile. Besides,
the type of id of CollectionsOnIndexProfiles has been changed to uuid,
which may be used, for example, as namespace for pinecone.