refactor: Support to upload files inner both tauri and browser
Known issue: when running in browser, the file content currently
will be uploaded to the backend in number array, which is very
inefficient in both data transfer and deserialization. It is
because the data transfer between frontend and tauri backend
must be serialized in json. Try to refactor this later.